Add test for Comments rest call

+++ libs/fluxc/src/main/java/org/wordpress/android/fluxc/model/stats/CommentsModel.kt
@@ -4,6 +4,6 @@ data class CommentsModel(
     var posts: List<Post>,
     var authors: List<Author>
 ) {
-    data class Post(val id: Int, val name: String, val comments: Int, val link: String)
+    data class Post(val id: Long, val name: String, val comments: Int, val link: String)
     data class Author(val name: String, val comments: Int, val link: String, val gravatar: String)
 }
